[KotlinTest] `ClassesFilter.groovy` to Kotlin
Part of #58
